rust: suggestions from clippy.

+++ src/apps/bench/rustnetbenchs/src/budp.rs
@@ -100,7 +100,7 @@ fn latency(t: &mut dyn WvTester) {
             ) {
                 let stop = TimeInstant::now();
 
-                wv_assert_eq!(t, *pkt_size, recv_size as usize);
+                wv_assert_eq!(t, *pkt_size, recv_size);
                 res.push(stop.duration_since(start));
             }
         }
@@ -191,7 +191,7 @@ fn bandwidth(t: &mut dyn WvTester) {
                     failures += 1;
                 },
                 Ok(size) => {
-                    received_bytes += size as usize;
+                    received_bytes += size;
                     receive_count += 1;
                     timeout = TimeInstant::now() + TIMEOUT;
                     failures = 0;
