+++ src/apps/rustnettests/src/tudp.rs
@@ -136,7 +136,7 @@ fn data(t: &mut dyn WvTester) {
                 &mut recv_buf,
                 TIMEOUT,
             ) {
-                wv_assert_eq!(t, *pkt_size, recv_size as usize);
+                wv_assert_eq!(t, *pkt_size, recv_size);
                 wv_assert_eq!(t, src, dest);
                 wv_assert_eq!(t, &recv_buf[0..recv_size], &send_buf[0..recv_size]);
                 break;
