+++ src/libs/rust/m3impl/src/server/server.rs
@@ -51,7 +51,7 @@ impl<'d> CapExchange<'d> {
     /// Creates a new `CapExchange` object, taking input arguments from `input` and putting output
     /// arguments into `output`.
     pub fn new(input: &'d ExchangeData, output: &'d mut ExchangeData) -> Self {
-        let len = (input.args.bytes as usize + 7) / 8;
+        let len = (input.args.bytes + 7) / 8;
         Self {
             src: M3Deserializer::new(&input.args.data[..len]),
             sink: M3Serializer::new(SliceSink::new(&mut output.args.data)),
